server/requirements.txt
PyYAML==6.0
requests==2.28.1
websockets==10.4
